refactor: types.go, unexport rcon
